Remove old PDA construction
